backend/ai/satellite_ai.py
"""
Satellite AI - Vision Transformer (ViT) for anomaly detection
Simulated implementation for demo purposes
"""

import numpy as np
from PIL import Image
from typing import Dict, List, Tuple, Optional
import random
from datetime import datetime

class SatelliteAI:
    """
    Vision Transformer model for detecting anomalies in satellite imagery
    Detects: fires, floods, land changes, cloud anomalies
    """
    
    def __init__(self):
        self.model_loaded = True
        self.confidence_threshold = 0.6
        self.detection_types = ["fire", "flood", "land_change", "cloud_anomaly"]
    
    def detect_anomalies(self, image_path: Optional[str] = None, image_array: Optional[np.ndarray] = None) -> List[Dict]:
        """
        Detect anomalies in satellite imagery
        
        Args:
            image_path: Path to satellite image file
            image_array: Numpy array of image (224x224x3)
        
        Returns:
            List of detected anomalies with confidence scores
        """
        # Simulated detection - in production, run actual ViT model
        num_detections = random.randint(0, 2)
        anomalies = []
        
        for i in range(num_detections):
            anomaly_type = random.choice(self.detection_types)
            confidence = random.uniform(0.65, 0.95)
            
            if confidence >= self.confidence_threshold:
                anomalies.append({
                    "type": anomaly_type,
                    "confidence": round(confidence, 3),
                    "bbox": {
                        "x1": random.randint(0, 150),
                        "y1": random.randint(0, 150),
                        "x2": random.randint(150, 224),
                        "y2": random.randint(150, 224)
                    },
                    "severity": self._calculate_severity(anomaly_type, confidence),
                    "detected_at": datetime.now().isoformat()
                })
        
        return anomalies
    
    def _calculate_severity(self, anomaly_type: str, confidence: float) -> float:
        """Calculate severity score based on anomaly type and confidence"""
        base_severity = {
            "fire": 0.8,
            "flood": 0.7,
            "land_change": 0.5,
            "cloud_anomaly": 0.4
        }
        
        base = base_severity.get(anomaly_type, 0.5)
        # Adjust based on confidence
        severity = base * confidence
        return round(min(severity, 1.0), 2)
    
    def classify_disaster(self, image_path: Optional[str] = None) -> Dict:
        """
        Classify the type of disaster from satellite image
        """
        anomalies = self.detect_anomalies(image_path)
        
        if not anomalies:
            return {
                "disaster_type": "none",
                "confidence": 0.0,
                "anomalies": []
            }
        
        # Get highest confidence anomaly
        primary_anomaly = max(anomalies, key=lambda x: x["confidence"])
        
        # Map anomaly types to disaster types
        disaster_map = {
            "fire": "fire",
            "flood": "flood",
            "land_change": "earthquake",  # or landslide
            "cloud_anomaly": "cyclone"
        }
        
        disaster_type = disaster_map.get(primary_anomaly["type"], "unknown")
        
        return {
            "disaster_type": disaster_type,
            "confidence": primary_anomaly["confidence"],
            "severity": primary_anomaly["severity"],
            "anomalies": anomalies,
            "detected_at": datetime.now().isoformat()
        }
    
    def process_tile(self, tile_path: str, location: Dict) -> Dict:
        """
        Process a satellite tile and return detection results with location
        """
        classification = self.classify_disaster(tile_path)
        
        return {
            **classification,
            "location": location,
            "tile_path": tile_path
        }

# Global instance
satellite_ai = SatelliteAI()



