backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
networkx==3.2.1
numpy==1.24.3
torch==2.1.0
opencv-python==4.8.1.78
pillow==10.1.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
websockets==12.0
aiofiles==23.2.1
